Changed size type to uint32

+++ Client/src/Network/AudioPacket.hpp
@@ -8,7 +8,7 @@
 namespace Babel {
     struct AudioPacket {
         std::uint64_t timestamp {};
-        std::uint64_t size {};
+        std::int32_t size {};
         unsigned char data[4000] {0};
     };
 }
